requirements.txt
kopf>=1.36.0
kubernetes>=26.1.0
gitpython>=3.1.0
asyncssh>=2.14.0
pyyaml>=6.0
asyncio
